secure_notes/requirements.txt
cryptography==44.0.2
python-dotenv==1.1.0 